+++ components/Landing/index.js
@@ -1,6 +1,5 @@
 import React from "react";
 import { Redirect, Link } from 'react-router-dom';
-import dynamic from 'next/dynamic';
 
 import Grid from '@material-ui/core/Grid';
 import Card from '@material-ui/core/Card';
@@ -11,26 +10,21 @@ import ListItem from '@material-ui/core/ListItem';
 import ListItemIcon from '@material-ui/core/ListItemIcon';
 import ListItemText from '@material-ui/core/ListItemText';
 import NearMe from '@material-ui/icons/NearMe';
-import CloudUpload from '@material-ui/icons/CloudUpload';
 import Pageview from '@material-ui/icons/Pageview';
 import FindReplace from '@material-ui/icons/FindReplace';
 import Public from '@material-ui/icons/Public';
-import Button from '@material-ui/core/Button';
 import { withStyles } from '@material-ui/core/styles';
 
 import { landingStyle,
          extraComponentStyle } from '../../styles/jss/theme.js'
 
 import SignatureSearch from '../SignatureSearch';
-import GenesetSearchBox from "./GenesetSearchBox";
 
 import { fetch_meta } from '../../util/fetch/meta'
 
-import { CurrentVersion, ListItemLink} from './Misc'
+import { CurrentVersion, ListItemLink, TitleCard} from './Misc'
 import { Charts, Stat } from "../Admin/dashboard.js";
 import { base_scheme as meta_base_scheme, base_url as meta_base_url } from "../../util/fetch/meta";
-
-const SearchBox = dynamic(() => import('../../components/MetadataSearch/SearchBox'))
 
 export default withStyles(landingStyle)(class extends React.Component {
 
@@ -132,6 +126,8 @@ export default withStyles(landingStyle)(class extends React.Component {
     )
   })
 
+  
+
   render(){
     const {classes} = this.props
     return(
@@ -141,56 +137,11 @@ export default withStyles(landingStyle)(class extends React.Component {
               alignItems={"center"}
               direction={"column"}>
           <Grid item xs={12}>
-            <Card className={classes.topCard}>
-              <Grid container 
-                spacing={24}
-                align="center"
-                justify="center">
-                <Grid item xs={12} md={7}>
-                  <Grid container 
-                    spacing={24}
-                    direction={"column"}
-                    align="center"
-                    justify="center">
-                    <Grid item xs={12}>
-                      <Typography variant="headline" className={classes.title} component="h3">
-                          SIGNATURE COMMONS
-                      </Typography>
-                      <Typography className={classes.subtitle} color="textSecondary">
-                          Search over half a million signatures collated from an ever-growing number of resources.
-                      </Typography>
-                    </Grid>
-                    <Grid item xs={12}>
-                      <SearchBox
-                        search={this.state.search}
-                        searchChange={this.searchChange}
-                      />
-                    </Grid>
-                    <Grid item xs={12}>
-                    {this.props.LibraryNumber > 0 && this.props.SignatureNumber > 0 && this.props.resource_signatures!==null ? 
-                        <CurrentVersion libraries_count={this.props.LibraryNumber}
-                                        signatures_count={this.props.SignatureNumber}
-                                        resources_count={Object.keys(this.props.resource_signatures).length}/>: null}
-                    </Grid>
-                    <Grid item xs={12}>
-                      <Link
-                        to={{ pathname: '/UploadCollection' }}
-                      >
-                        <Button variant="contained" color="secondary" className={classes.button}>
-                          Upload your signatures
-                          <CloudUpload className={classes.rightIcon}/>
-                        </Button>
-                      </Link>
-                    </Grid>
-                  </Grid>
-                </Grid>
-                <Grid item xs={12} md={5}>
-                  <Grid item xs={12}>
-                    {this.geneset_searchbox(this.props)}
-                  </Grid>
-                </Grid>
-              </Grid>
-            </Card>
+            <TitleCard search={this.state.search}
+                       searchChange={this.searchChange}
+                       type={this.state.type}
+                       submit={this.submit}
+                       {...this.props} />
           </Grid>
           <Grid item xs={12}>
             <Grid container
